Offline Version
For grading purposes, the file should run from a local filesystem. Content was pasted directly into the HTML instead of using AJAX to load. I was testing using Node.js which is why is was working for me. I will be sure to download the zip from github and run locally from now on.